+++ R/fsle.R
@@ -84,8 +84,7 @@ fsle <- function(env_dat, u = "UO", v = "VO", final_separation = 50,
   }
 
   steps <- time_steps(env_dat)
-  step_times <- as.numeric(as.Date(paste(steps$YEAR, steps$MONTH, steps$DAY,
-                                          sep = "-")))
+  step_times <- step_time_days(steps)
   velocity <- lapply(seq_len(nrow(steps)), function(i) {
     rasterize_step(env_dat[step_rows(env_dat, steps[i, ]), ], c(u, v))
   })