+++ R/ftle.R
@@ -90,8 +90,7 @@ ftle <- function(env_dat, u = "UO", v = "VO", integration_days = 14,
   }
 
   steps <- time_steps(env_dat)
-  step_times <- as.numeric(as.Date(paste(steps$YEAR, steps$MONTH, steps$DAY,
-                                          sep = "-")))
+  step_times <- step_time_days(steps)
   velocity <- lapply(seq_len(nrow(steps)), function(i) {
     rasterize_step(env_dat[step_rows(env_dat, steps[i, ]), ], c(u, v))
   })